@@ -1,82 +1,61 @@
 {
   "name": "wes-bos-blog",
-  "description": "Wesssss Site",
+  "description": "Wes Bos Site",
   "version": "1.0.0",
   "author": "Wes Bos <wesbos@gmail.com>",
   "engines": {
-    "node": ">=12.14.0"
+    "node": ">=18.0.0"
   },
   "dependencies": {
-    "@mdx-js/loader": "1.6.22",
-    "@mdx-js/mdx": "1.6.22",
-    "@mdx-js/react": "^1.6.22",
-    "core-js": "^3.12.1",
-    "date-fns": "^2.21.3",
-    "dotenv": "^9.0.2",
-    "esm": "^3.2.25",
-    "file-type": "^16.4.0",
-    "gatsby": "^3.5.1",
-    "gatsby-image": "^3.5.0",
+    "@mdx-js/react": "^2.2.1",
+    "date-fns": "^2.29.3",
+    "dotenv": "^16.0.3",
+    "file-type": "^18.0.0",
+    "gatsby": "^5.4.0",
     "gatsby-parallel-runner": "^1.3.0",
-    "gatsby-plugin-feed": "^3.5.0",
-    "gatsby-plugin-google-analytics": "^3.5.0",
-    "gatsby-plugin-image": "^1.5.0",
-    "gatsby-plugin-layout": "^2.5.0",
-    "gatsby-plugin-manifest": "^3.5.0",
-    "gatsby-plugin-mdx": "^2.5.1",
-    "gatsby-plugin-nprogress": "^3.5.0",
-    "gatsby-plugin-offline": "^4.5.1",
+    "gatsby-plugin-google-analytics": "^5.4.0",
+    "gatsby-plugin-image": "^3.4.0",
+    "gatsby-plugin-layout": "^4.4.0",
+    "gatsby-plugin-mdx": "^5.4.0",
     "gatsby-plugin-prettier-build": "^0.4.4",
-    "gatsby-plugin-react-helmet": "^4.5.0",
-    "gatsby-plugin-remove-trailing-slashes": "^3.5.0",
-    "gatsby-plugin-sharp": "^3.5.0",
+    "gatsby-plugin-react-helmet": "^6.4.0",
+    "gatsby-plugin-sharp": "^5.4.0",
     "gatsby-plugin-snipcartv3": "^2.3.3",
-    "gatsby-plugin-styled-components": "^4.5.0",
-    "gatsby-plugin-twitter": "^3.5.0",
-    "gatsby-remark-autolink-headers": "^4.2.0",
-    "gatsby-remark-copy-linked-files": "^4.2.1",
-    "gatsby-remark-embedder": "^5.0.0",
-    "gatsby-remark-images": "^5.2.1",
-    "gatsby-remark-responsive-iframe": "^4.2.1",
-    "gatsby-remark-shiki": "^0.2.0",
-    "gatsby-remark-smartypants": "^4.2.0",
-    "gatsby-remark-vscode": "^3.2.1",
-    "gatsby-source-filesystem": "^3.5.0",
-    "gatsby-transformer-remark": "^4.2.0",
-    "gatsby-transformer-sharp": "^3.5.0",
-    "html-entities": "^2.3.2",
+    "gatsby-plugin-styled-components": "^6.4.0",
+    "gatsby-remark-autolink-headers": "^6.4.0",
+    "gatsby-remark-copy-linked-files": "^6.4.0",
+    "gatsby-remark-embedder": "^6.0.1",
+    "gatsby-remark-images": "^7.4.0",
+    "gatsby-remark-responsive-iframe": "^6.4.0",
+    "gatsby-remark-smartypants": "^6.4.0",
+    "gatsby-remark-vscode": "^3.3.1",
+    "gatsby-source-filesystem": "^5.4.0",
+    "gatsby-transformer-remark": "^6.4.0",
+    "gatsby-transformer-sharp": "^5.4.0",
+    "github-slugger": "^2.0.0",
     "isomorphic-fetch": "^3.0.0",
-    "netlify-lambda": "^2.0.7",
+    "netlify-lambda": "^2.0.15",
+    "netlify-plugin-gatsby-cache": "^0.3.2",
     "normalize.css": "^8.0.1",
-    "prism-react-renderer": "^1.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-icons": "^4.2.0",
-    "react-youtube": "^7.13.1",
-    "remark-autolink-headings": "^6.0.1",
+    "react-icons": "^4.7.1",
     "resize-observer-polyfill": "^1.5.1",
     "speakingurl": "^14.0.1",
-    "string.prototype.replaceall": "^1.0.5",
-    "styled-components": "^5.3.0",
-    "theme-cobalt2": "^2.1.6"
+    "styled-components": "^5.3.6",
+    "theme-cobalt2": "^2.1.6",
+    "waait": "^1.0.5"
   },
   "devDependencies": {
+    "@types/node": "^18.11.18",
+    "@types/react-helmet": "^6.1.6",
+    "@types/styled-components": "^5.1.26",
     "babel-eslint": "^10.1.0",
-    "eslint": "^7.26.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-config-wesbos": "1.0.1",
-    "eslint-plugin-html": "^6.1.2",
-    "eslint-plugin-import": "^2.23.2",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.23.2",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint": "^8.31.0",
+    "eslint-config-wesbos": "^3.2.0",
     "expand-url": "^0.1.3",
-    "gh-pages": "^3.1.0",
-    "glob": "^7.1.7",
-    "prettier": "^2.3.0"
+    "prettier": "^2.8.2"
   },
   "homepage": "https://wesbos.com",
   "keywords": [
