@@ -4,23 +4,12 @@ export default function LatestCourse() {
   return (
     <div>
       <h3>
-        <span className="highlight">Master Gatsby</span>
+        <span className="highlight">Beginner JavaScript</span>
       </h3>
-      <img
-        src="https://res.cloudinary.com/wesbos/image/fetch/w_700,q_auto,f_auto/https://courses.wesbos.com/images/GAT/GAT-social-share.png"
-        alt="Master Gatsby"
-      />
-      <p>
-        Building modern websites is tough. Preloading, routing, compression,
-        critical CSS, caching, scaling and bundlers all make for blazing fast
-        websites, but extra development and tooling get in the way.
-      </p>
-      <p>
-        Gatsby is a React.js framework that does it all for you. This course
-        will teach you how to build your websites and let Gatsby take care of
-        all the Hard Stuff™.
-      </p>
-      <a href="https://MasterGatsby.com">MasterGatsby.com</a>
+      <img src="https://res.cloudinary.com/wesbos/image/fetch/w_400,q_auto,f_auto/https://courses.wesbos.com/images/BJS/BJS-Social-Share.png" alt="Beginner JavaScript" />
+      <p>A fun, exercise heavy approach to learning Modern JavaScript from scratch. This is a course for absolute beginners or anyone looking to brush up on their fundamentals. Start here if you are new to JS or programming in general!</p>
+
+      <a href="https://BeginnerJavaScript.com">BeginnerJavaScript.com</a>
     </div>
   );
 }
