@@ -14,25 +14,28 @@ function getBaseURL() {
 const baseURL = getBaseURL();
 
 export function PostMetaTags({ post }) {
-  const canonical = pathJoin('https://wesbos.com', post.frontmatter.slug);
-  const url = pathJoin(baseURL, post.frontmatter.slug);
+  // TODO: We should be using post.fields.slug, but in some pages it's passed in as "frontmatter". It should be fixed
+  const slug = post.fields?.slug || post.frontmatter?.slug;
+  if (!slug) {
+    console.log('No slug found for post', { post });
+    throw new Error('No slug found for post', { post });
+  }
+  const canonical = pathJoin('https://wesbos.com', slug);
+  const url = pathJoin(baseURL, slug);
   const thumbnailData = {
     title: post.frontmatter.title,
     url,
     thumbnail: post.frontmatter.image?.publicURL,
   };
-  const thumbnailQuery = new URLSearchParams(
-    Object.fromEntries(
-      Object.entries(thumbnailData).filter(([key, val]) => val !== undefined)
-    )
-  ).toString();
+  /* eslint no-unused-vars: ["error", { "destructuredArrayIgnorePattern": "^_" }] */
+  const thumbnailQuery = new URLSearchParams(Object.fromEntries(Object.entries(thumbnailData).filter(([_key, val]) => val !== undefined))).toString();
 
   const ogImage = `${baseURL}/.netlify/functions/ogimage?${thumbnailQuery}`;
   return (
     <Helmet>
-      <link rel="canonical" href={canonical} />
       <meta name="generator" content="Wes Bos on Gatsby!" />
-      <meta name="twitter:card" content="summary" />
+      <link rel="canonical" href={canonical} />
+      <meta name="twitter:card" content="summary_large_image" />
       <meta name="twitter:site" content="@wesbos" />
       <meta name="twitter:creator" content="@wesbos" />
       <meta name="twitter:url" content={url} />
@@ -43,43 +46,37 @@ export function PostMetaTags({ post }) {
       <meta property="og:title" content={post.frontmatter.title} />
       <meta property="og:url" content={url} />
       <meta property="og:description" content={post.excerpt} />
-      {post.frontmatter.date ? (
-        <meta
-          property="article:published_time"
-          content={new Date(post.frontmatter.date).toISOString()}
-        />
-      ) : null}
+      {post.frontmatter.date ? <meta property="article:published_time" content={new Date(post.frontmatter.date).toISOString()} /> : null}
 
       <meta property="og:site_name" content="Wes Bos" />
       <meta property="og:image" content={ogImage} />
       <meta property="og:image:width" content="1200" />
       <meta property="og:image:height" content="630" />
       <meta property="og:locale" content="en_CA" />
-      <title>{post.frontmatter.title} - Wes Bos</title>
+      <title>{`${post.frontmatter.title} - Wes Bos`}</title>
     </Helmet>
   );
 }
 
 export function TipsMetaTags({ post }) {
-  const canonical = `https://wesbos.com/tip/${post.frontmatter.slug}`;
-  const url = `${baseURL}/tip/${post.frontmatter.slug}`;
+  const slug = post.fields?.slug || post.frontmatter?.slug;
+  const canonical = pathJoin('https://wesbos.com/tip/', slug);
+  const url = pathJoin(baseURL, '/tip/', slug);
   const thumbnailData = {
     title: post.excerpt,
     url,
     thumbnail: post.frontmatter.images?.[0]?.publicURL,
+    cache: 'busta',
   };
-  const thumbnailQuery = new URLSearchParams(
-    Object.fromEntries(
-      Object.entries(thumbnailData).filter(([key, val]) => val !== undefined)
-    )
-  ).toString();
+  /* eslint no-unused-vars: ["error", { "destructuredArrayIgnorePattern": "^_" }] */
+  const thumbnailQuery = new URLSearchParams(Object.fromEntries(Object.entries(thumbnailData).filter(([_key, val]) => val !== undefined))).toString();
 
   const ogImage = `${baseURL}/.netlify/functions/ogimage?${thumbnailQuery}`;
   return (
     <Helmet>
       <link rel="canonical" href={canonical} />
       <meta name="generator" content="Wes Bos on Gatsby!" />
-      <meta name="twitter:card" content="summary" />
+      <meta name="twitter:card" content="summary_large_image" />
       <meta name="twitter:site" content="@wesbos" />
       <meta name="twitter:creator" content="@wesbos" />
       <meta name="twitter:url" content={url} />
@@ -90,16 +87,13 @@ export function TipsMetaTags({ post }) {
       <meta property="og:title" content={post.frontmatter.title} />
       <meta property="og:url" content={url} />
       <meta property="og:description" content={post.excerpt} />
-      <meta
-        property="article:published_time"
-        content={new Date(post.frontmatter.date).toISOString()}
-      />
+      <meta property="article:published_time" content={new Date(post.frontmatter.date).toISOString()} />
       <meta property="og:site_name" content="Wes Bos" />
       <meta property="og:image" content={ogImage} />
       <meta property="og:image:width" content="1200" />
       <meta property="og:image:height" content="630" />
       <meta property="og:locale" content="en_CA" />
-      <title>{post.frontmatter.title} - Wes Bos</title>
+      <title>{`${post.frontmatter.title} - Wes Bos`}</title>
     </Helmet>
   );
 }
