@@ -12,14 +12,12 @@ export default function useImage() {
    * @returns string[]
    */
   const { images } = useStaticQuery(graphql`
-    query {
+    {
       images: allFile(filter: { dir: { regex: "/assets/images/" } }) {
         edges {
           node {
             childImageSharp {
-              fluid(maxWidth: 1000) {
-                src
-              }
+              gatsbyImageData(placeholder: BLURRED, layout: FULL_WIDTH)
             }
           }
         }
@@ -28,9 +26,7 @@ export default function useImage() {
   `);
   return {
     getImagePath(fileName) {
-      return images.edges
-        .map((x) => x.node.childImageSharp.fluid.src)
-        .find((path) => path.endsWith(fileName));
+      return images.edges.map((x) => x.node.childImageSharp.gatsbyImageData.src).find((path) => path.endsWith(fileName));
     },
   };
 }
