@@ -2,59 +2,50 @@ import React from 'react';
 import { Link, graphql } from 'gatsby';
 import format from 'date-fns/format';
 import Img from '../components/Img';
-import PostGrid, { PostGridItem } from '../styles/PostGrid';
 import H from '../components/mdxComponents/Headings';
 import Pagination from '../components/Pagination';
 import { PostMetaTags } from '../components/MetaTags';
+import PostGrid, { PostGridItem } from '../assets/styles/PostGrid';
 
-const Blog = function ({ data, pageContext, path }) {
+export default function Blog({ data, pageContext, location }) {
   if (!data) return <p>Shooooot! No Post found!</p>;
   return (
     <>
-      <Pagination
-        currentPage={pageContext.currentPage}
-        totalCount={data.allMdx.totalCount}
-        pathPrefix="/blog/"
-      />
+      <Pagination currentPage={pageContext.currentPage} totalCount={data.allMdx.totalCount} pathPrefix="/blog/" />
       <PostMetaTags
         post={{
+          fields: {
+            slug: location.pathname, // TODO: TEST
+          },
           frontmatter: {
-            slug: path,
-            title: `Blog ${
-              pageContext.currentPage ? `- Page ${pageContext.currentPage}` : ''
-            }`,
+            slug: location.pathname,
+            title: `Blog ${pageContext.currentPage ? `- Page ${pageContext.currentPage}` : ''}`,
           },
         }}
       />
       <PostGrid>
         {data.allMdx &&
-          data.allMdx.edges.map(function ({ node: post }) {
+          data.allMdx.edges.map(({ node: post }) => {
             const hasImage = !!post.frontmatter.image;
             return (
               <PostGridItem key={post.id} hasImage={hasImage}>
-                {post.frontmatter.image &&
-                  post.frontmatter.image.childImageSharp && (
-                    <div>
-                      {hasImage && (
-                        <Link to={post.fields.slug}>
-                          <Img image={post.frontmatter.image} />
-                        </Link>
-                      )}
-                      <div className="postMeta">
-                        <time dateTime={post.frontmatter.date}>
-                          {format(
-                            new Date(post.frontmatter.date),
-                            'MMMM d, yyyy'
-                          )}
-                        </time>
-                        <ul className="categories">
-                          {post.frontmatter.category.map((cat) => (
-                            <li key={cat}>{cat}</li>
-                          ))}
-                        </ul>
-                      </div>
+                {post.frontmatter.image && post.frontmatter.image.childImageSharp && (
+                  <div>
+                    {hasImage && (
+                      <Link to={post.fields.slug}>
+                        <Img image={post.frontmatter.image} />
+                      </Link>
+                    )}
+                    <div className="postMeta">
+                      <time dateTime={post.frontmatter.date}>{format(new Date(post.frontmatter.date), 'MMMM d, yyyy')}</time>
+                      <ul className="categories">
+                        {post.frontmatter.category.map((cat) => (
+                          <li key={cat}>{cat}</li>
+                        ))}
+                      </ul>
                     </div>
-                  )}
+                  </div>
+                )}
                 <div>
                   <H as="h3">
                     <Link to={post.fields.slug}>{post.frontmatter.title}</Link>
@@ -65,16 +56,10 @@ const Blog = function ({ data, pageContext, path }) {
             );
           })}
       </PostGrid>
-      <Pagination
-        currentPage={pageContext.currentPage}
-        totalCount={data.allMdx.totalCount}
-        pathPrefix="/blog/"
-      />
+      <Pagination currentPage={pageContext.currentPage} totalCount={data.allMdx.totalCount} pathPrefix="/blog/" />
     </>
   );
-};
-
-export default Blog;
+}
 
 export const pageQuery = graphql`
   query blogPosts($skip: Int! = 0) {
@@ -84,12 +69,7 @@ export const pageQuery = graphql`
         description
       }
     }
-    allMdx(
-      filter: { fields: { collection: { eq: "post" } } }
-      sort: { fields: [frontmatter___date], order: DESC }
-      limit: 10
-      skip: $skip
-    ) {
+    allMdx(filter: { fields: { collection: { eq: "post" } } }, sort: { frontmatter: { date: DESC } }, limit: 10, skip: $skip) {
       totalCount
       edges {
         node {
@@ -102,6 +82,7 @@ export const pageQuery = graphql`
           frontmatter {
             title
             date
+            slug
             category
             image {
               ...ImageFields
