@@ -1,6 +1,5 @@
 import React from 'react';
 import { graphql } from 'gatsby';
-import { MDXRenderer } from 'gatsby-plugin-mdx';
 import { IoLogoGithub } from 'react-icons/io';
 import { Helmet } from 'react-helmet';
 import H from '../components/mdxComponents/Headings';
@@ -15,7 +14,7 @@ import { useActiveId } from '../hooks/useActiveId';
 import { BeginnerJavaScript } from '../components/beginnerJavaScript';
 
 export const pageQuery = graphql`
-  query($slug: String!) {
+  query ($slug: String!) {
     site {
       siteMetadata {
         title
@@ -24,7 +23,11 @@ export const pageQuery = graphql`
     }
     mdx(fields: { slug: { eq: $slug } }) {
       id
-      fileAbsolutePath
+      parent {
+        ... on File {
+          absolutePath
+        }
+      }
       frontmatter {
         title
         slug
@@ -36,16 +39,14 @@ export const pageQuery = graphql`
   }
 `;
 
-function JavaScriptNotesTemplate({ data: { mdx: post }, scope, pageContext }) {
+function JavaScriptNotesTemplate({ data: { mdx: post }, scope, pageContext, children }) {
   const activeId = useActiveId(getIds(post.tableOfContents.items));
 
   if (!post) {
     return <p>No Post Found? This should be a 404</p>;
   }
 
-  const editURL = `https://github.com/wesbos/wesbos/tree/master/src/${
-    post.fileAbsolutePath.split('/src/')[1]
-  }`;
+  const editURL = `https://github.com/wesbos/wesbos/tree/master/src/${post.parent.absolutePath.split('/src/')[1]}`;
   return (
     <JavaScriptNotesStyles>
       <TableOfContents activeId={activeId} currentPage={pageContext.slug} />
@@ -53,7 +54,7 @@ function JavaScriptNotesTemplate({ data: { mdx: post }, scope, pageContext }) {
         <PostHeaderStyles>
           <PostMetaTags post={post} />
           <H>{post.frontmatter.title}</H>
-          <BeginnerJavaScript/>
+          <BeginnerJavaScript />
           <div className="postMeta">
             <span>{post.frontmatter.category.join(', ')}</span>
             <a rel="noopener noreferrer" target="_blank" href={editURL}>
@@ -61,37 +62,19 @@ function JavaScriptNotesTemplate({ data: { mdx: post }, scope, pageContext }) {
             </a>
           </div>
         </PostHeaderStyles>
-        <MDXRenderer
-          scope={{
-            ...scope,
-          }}
-        >
-          {post.body}
-        </MDXRenderer>
+        {children}
         <EditDialogStyles>
-          <p>
-            Find an issue with this post? Think you could clarify, update or add
-            something?
-          </p>
-          <p>
-            All my posts are available to edit on Github. Any fix, little or
-            small, is appreciated!
-          </p>
+          <p>Find an issue with this post? Think you could clarify, update or add something?</p>
+          <p>All my posts are available to edit on Github. Any fix, little or small, is appreciated!</p>
           <p>
             <a rel="noopener noreferrer" target="_blank" href={editURL}>
               <IoLogoGithub /> Edit on Github
             </a>
           </p>
         </EditDialogStyles>
-        <ContentNav
-          pathPrefix={pageContext.pathPrefix}
-          prev={pageContext.prev}
-          next={pageContext.next}
-        />
+        <ContentNav pathPrefix={pageContext.pathPrefix} prev={pageContext.prev} next={pageContext.next} />
         <Helmet>
-          <title>
-            {post.frontmatter.title} - Beginner JavaScript - Wes Bos
-          </title>
+          <title>{post.frontmatter.title} - Beginner JavaScript - Wes Bos</title>
         </Helmet>
       </article>
     </JavaScriptNotesStyles>
