@@ -1,8 +1,6 @@
 import React from 'react';
 import { graphql } from 'gatsby';
-import { MDXRenderer } from 'gatsby-plugin-mdx';
 import Img from '../components/Img';
-import Layout from '../components/Layout';
 import ContentNav from '../components/ContentNav';
 import H from '../components/mdxComponents/Headings';
 import TipMeta from '../components/TipMeta';
@@ -10,7 +8,7 @@ import { Grid } from '../components/styles/Grid';
 import { TipsMetaTags } from '../components/MetaTags';
 
 export const pageQuery = graphql`
-  query($slug: String!) {
+  query ($slug: String!) {
     site {
       siteMetadata {
         title
@@ -20,6 +18,9 @@ export const pageQuery = graphql`
     mdx(fields: { slug: { eq: $slug } }) {
       id
       excerpt
+      fields {
+        slug
+      }
       frontmatter {
         title
         slug
@@ -35,7 +36,8 @@ export const pageQuery = graphql`
   }
 `;
 
-export default function TipTemplate({ data: { mdx: tip }, pageContext }) {
+/* eslint-disable jsx-a11y/media-has-caption */
+export default function TipTemplate({ data: { mdx: tip }, pageContext, children }) {
   return (
     <>
       <div>
@@ -44,21 +46,11 @@ export default function TipTemplate({ data: { mdx: tip }, pageContext }) {
           <H>Hot Tip</H>
           <TipMeta tip={tip} />
         </Grid>
-        <MDXRenderer>{tip.body}</MDXRenderer>
-        {tip.frontmatter.videos &&
-          tip.frontmatter.videos.map(url => (
-            <video key={url} src={url} autoPlay mute loop />
-          ))}
-        {tip.frontmatter.images &&
-          tip.frontmatter.images.map(image => (
-            <Img key={image.id} image={image} />
-          ))}
+        {children}
+        {tip.frontmatter.videos && tip.frontmatter.videos.map((url) => <video key={url} src={url} autoPlay muted loop />)}
+        {tip.frontmatter.images && tip.frontmatter.images.map((image) => <Img key={image.id} image={image} alt={tip.body} />)}
       </div>
-      <ContentNav
-        pathPrefix={pageContext.pathPrefix}
-        prev={pageContext.prev}
-        next={pageContext.next}
-      />
+      <ContentNav pathPrefix={pageContext.pathPrefix} prev={pageContext.prev} next={pageContext.next} />
     </>
   );
 }
